backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-json-logger==3.2.1
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-exporter-prometheus==0.57b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-instrumentation-logging==0.57b0
httpx==0.28.1
